Split tructs in different files in modules dir
