+++ src/models/cli.rs
@@ -1,0 +1,10 @@
+use clap::Parser;
+
+/* -------------------------- APP PARAMS PARSER -------------------------- */
+#[derive(Parser)]
+pub struct Cli {
+    pub command: String,
+    pub value: u8,
+}
+
+impl Cli {}
